backend/requirements.txt
flask
flask-cors
requests
pillow
pytesseract
